+++ tests/testthat/helper_test_pipeops.R
@@ -15,6 +15,7 @@ PipeOpDebugBasic = R6Class("PipeOpDebugBasic",
     .predict = function(inputs) {
       catf("Predicting %s", self$id)
       self$state = c(self$state, inputs)
+      inputs
     }
   )
 )
@@ -57,7 +58,7 @@ PipeOpDebugMulti = R6Class("PipeOpDebugMulti",
     },
     .predict = function(inputs) {
       catf("Predicting %s with input %s and state %s",
-        self$id, deparse_list_safe(inputs), deparse_list_safe(self$state))
+        self$id, deparse_list_safe(inputs), deparse_list_safe(self$state[-which(names(self$state) %in% c("log", "train_time"))]))
       iin = inputs[[1]]
       as.list(iin + seq_len(self$nout))
     }
@@ -84,3 +85,63 @@ VarargPipeop = R6Class("VarargPipeop",
     }
   )
 )
+
+PipeOpDebugEncapsulate = R6Class("PipeOpDebugEncapsulate",
+  inherit = PipeOp,
+  public = list(
+    initialize = function(id = "debug.encapsulate", param_vals = list()) {
+      param_set = ParamSet$new(list(
+        ParamLgl$new("message_train", default = FALSE, tags = "train"),
+        ParamLgl$new("message_predict", default = FALSE, tags = "predict"),
+        ParamLgl$new("warning_train", default = FALSE, tags = "train"),
+        ParamLgl$new("warning_predict", default = FALSE, tags = "predict"),
+        ParamLgl$new("error_train", default = FALSE, tags = "train"),
+        ParamLgl$new("error_predict", default = FALSE, tags = "predict"),
+        ParamLgl$new("segfault_train", default = FALSE, tags = "train"),
+        ParamLgl$new("segfault_predict", default = FALSE, tags = "predict"),
+        ParamLgl$new("predict_missing", default = FALSE, tags = "predict")
+      ))
+      super$initialize(id = id, param_set = param_set, param_vals = param_vals,
+        input = data.table(name = "input", train = "*", predict = "*"),
+        output = data.table(name = "output", train = "*", predict = "*")
+      )
+    }),
+  private = list(
+    .train = function(inputs) {
+      pv = self$param_set$get_values(tags = "train")
+
+      if (isTRUE(pv[["message_train"]])) {
+        message("Message from classif.debug->train()")
+      }
+      if (isTRUE(pv[["warning_train"]])) {
+        warning("Warning from classif.debug->train()")
+      }
+      if (isTRUE(pv[["error_train"]])) {
+        stop("Error from classif.debug->train()")
+      }
+      if (isTRUE(pv[["segfault_train"]])) {
+        get("attach")(structure(list(), class = "UserDefinedDatabase"))
+      }
+      self$state = inputs
+    },
+    .predict = function(inputs) {
+      pv = self$param_set$get_values(tags = "predict")
+
+      if (isTRUE(pv[["message_predict"]])) {
+        message("Message from classif.debug->predict()")
+      }
+      if (isTRUE(pv[["warning_predict"]])) {
+        warning("Warning from classif.debug->predict()")
+      }
+      if (isTRUE(pv[["error_predict"]])) {
+        stop("Error from classif.debug->predict()")
+      }
+      if (isTRUE(pv[["segfault_predict"]])) {
+        get("attach")(structure(list(), class = "UserDefinedDatabase"))
+      }
+      self$state = c(self$state, inputs)
+      inputs
+    }
+  )
+)
+