+++ tests/testthat/test_pipeop_proxy.R
@@ -5,7 +5,8 @@ test_that("PipeOpProxy - basic properties", {
   pop = PipeOpProxy$new(param_vals = list(content = PipeOpNOP$new()))
   expect_pipeop(pop)
   expect_equal(train_pipeop(pop, inputs = list(task))[[1L]], task)
-  expect_equal(pop$state, list(nop = list()))
+  expect_length(pop$state, 3L)
+  expect_length(pop$state$nop, 2L)
   expect_equal(predict_pipeop(pop, inputs = list(task))[[1L]], task)
 })
 
@@ -71,7 +72,7 @@ test_that("PipeOpProxy - Complicated Graphs", {
   # ---------
   # two parallel single-input-single-output pipeops
 
-  pop = po("proxy", content = list(PipeOpDebugMulti$new(1, 1, "debug1"), PipeOpDebugMulti$new(1, 1, "debug2")), outnum = 2)
+  pop = po("proxy", content = list(PipeOpDebugMulti$new(1, 1, "debug1"), PipeOpDebugMulti$new(1, 1, "debug2")), outnum = 2L)
 
   expect_output(
     expect_equal(pop$train(list(1)), list(output1 = 2, output2 = 2)),
@@ -97,15 +98,15 @@ Predicting debug2 with input list\\(input_1 = 4\\) and state list\\(input_1 = 2\
   # ---------
   # NOP | feature union | NOP; feature union has vararg
 
-  pop = po("proxy", content = list(PipeOpDebugMulti$new(1, 1, "debug1"), PipeOpFeatureUnion$new(), PipeOpDebugMulti$new(1, 1, "debug2")), outnum = 3)
+  pop = po("proxy", content = list(PipeOpDebugMulti$new(1, 1, "debug1"), PipeOpFeatureUnion$new(), PipeOpDebugMulti$new(1, 1, "debug2")), outnum = 3L)
 
   tsk1 = tsk("iris")
-  tsk2 = po("pca")$train(list(tsk1))[[1]]
-  tsk3 = po("ica")$train(list(tsk1))[[1]]
+  tsk2 = po("pca")$train(list(tsk1))[[1L]]
+  tsk3 = po("ica")$train(list(tsk1))[[1L]]
 
   expect_output(
       expect_equal(pop$train(list(1, tsk1, tsk2, tsk3, 2)),
-        list(output1 = 2, output2 = po("featureunion")$train(list(tsk1, tsk2, tsk3))[[1]], output3 = 3)),
+        list(output1 = 2, output2 = po("featureunion")$train(list(tsk1, tsk2, tsk3))[[1L]], output3 = 3)),
     "^Training debug1 with input list\\(input_1 = 1\\)
 Training debug2 with input list\\(input_1 = 2\\)$")